[fix]: 조건문 재확인 코드 제거
