+++ src/main/java/org/c4marathon/assignment/event/JoinEventDto.java
@@ -1,0 +1,12 @@
+package org.c4marathon.assignment.common.config;
+
+import org.c4marathon.assignment.user.domain.User;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class JoinEventDto {
+	private User user;
+}
